% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_contact_trends.R
\name{build_contact_trends}
\alias{build_contact_trends}
\title{Build Contact Trends}
\usage{
build_contact_trends(contact_data, role_data, type = "visits", first_fy = 2015)
}
\arguments{
\item{contact_data}{a data frame of contact reports, built by get_contact_reports(). Filter by date, fundraiser and unit, if desired, BEFORE using this function.}

\item{role_data}{a data frame of fundraisers in the desired role, built by get_roles(). Roles should be filtered before using this function.}

\item{type}{a character strings, the type of visits. 'visits', 'unique households' or 'significant contacts'}

\item{first_fy}{what fiscal year should the bar charts start at}
}
\value{
a dataframe to test against
}
\description{
build a data frame to test USM/CSM/DOM Tab #2 against.
}
