% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_qualifications_trend.R
\name{build_qualifications_trend}
\alias{build_qualifications_trend}
\title{Build Qualifications Trend Date}
\usage{
build_qualifications_trend(qual_data, first_fy = 2015)
}
\arguments{
\item{qual_data}{a data frame of qualification data, build by get_qualification_history(). Filter to get the relevant proposals BEFORE using this functino.}

\item{first_fy}{the first fiscal year that should appear on the bar chart}
}
\value{
a dataframe to test against
}
\description{
build a data frame to test USM/CSM/DOM Tab #2 against
}
