% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tab2.R
\name{build_tab2}
\alias{build_tab2}
\title{Build Tab 2}
\usage{
build_tab2(
  proposal_data = get_tab2_data()["proposal_data"],
  contact_data = get_tab2_data()["contact_data"],
  role_data = get_tab2_data()["role_data"],
  start_fy = 2015
)
}
\arguments{
\item{proposal_data}{a data frame of proposal data, as generated by get_proposal_data (or get_tab2_data())}

\item{contact_data}{a data frame of contact report data, as generated by get_contact_reports (or get_tab2_data())}

\item{role_data}{a data frame of role data, as generated by get_roles (or get_tab2_data())}

\item{start_fy}{what fiscal year should the bar charts start at}
}
\value{
a list with 8 elements, one for each of the metrics on Tab 2 (qualifications isn't done yet)
}
\description{
Build Tab 2
}
