% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_contacted_households.R
\name{count_contacted_households}
\alias{count_contacted_households}
\title{Count Households Contacted (ie. Unique Visits)}
\usage{
count_contacted_households(contact_data)
}
\arguments{
\item{contact_data}{a data frame of contact data built by get_contact_reports()}
}
\value{
all the data needed to verify the numbers on the USM/CSM/DOM Replacement Tab #2
}
\description{
for use in Tab2 build_contact_trends(). Not exported.
}
