% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_contacts.R
\name{count_contacts}
\alias{count_contacts}
\title{Count Contacts}
\usage{
count_contacts(contact_data)
}
\arguments{
\item{contact_data}{a data frame built by get_contact_reports()}
}
\value{
all the data needed to verify the numbers on the USM/CSM/DOM Replacement Tab #2
}
\description{
primarily for use in Tab 2 build_* functions. Given a local contact report dataframe (built from get_contact_reports() ), count contacts. To count visits, filter down to visits before using this function.
}
