% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_qualifications.R
\name{csm_contacts}
\alias{csm_contacts}
\title{Pull Contact Reports like the Campus Summary Metrics}
\usage{
csm_contacts(maindata_q, start_dt = "2021-07-01", end_dt = "2022-06-30")
}
\arguments{
\item{maindata_q}{a text string to pull query the maindata CTE, as produced by csm_maindata_details()}

\item{start_dt}{the activity start date, as a string}

\item{end_dt}{the activity end date, as a string}
}
\value{
a text string that can be used to query the CDW to get contact reports
}
\description{
Pull Contact Reports like the Campus Summary Metrics
}
