% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_target_asks.R
\name{csm_filter_to_target_asks}
\alias{csm_filter_to_target_asks}
\title{Filter a Portfolio down the the Target Asks, ie. Planned Asks}
\usage{
csm_filter_to_target_asks(portfolio_local)
}
\arguments{
\item{portfolio_local}{a collected, in local memory, data frame built from csm_build_portfolio_query()}
}
\value{
a list of proposal assignments that have target asks
}
\description{
Filter a Portfolio down the the Target Asks, ie. Planned Asks
}
