% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_actual_asks.R
\name{csm_get_fundraiser_actual_asks}
\alias{csm_get_fundraiser_actual_asks}
\title{CSM Actual Asks}
\usage{
csm_get_fundraiser_actual_asks(portfolio_q)
}
\arguments{
\item{portfolio_q}{a portfolio query built by csm_build_portfolio_query()}
}
\value{
a query string that can be used to pull the proposals that make up Actual Asks on the CSM
}
\description{
CSM Actual Asks
}
