% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_target_asks.R
\name{csm_get_fundraiser_target_asks}
\alias{csm_get_fundraiser_target_asks}
\title{Get a CSM-like fundraiser total}
\usage{
csm_get_fundraiser_target_asks(portfolio_q)
}
\arguments{
\item{portfolio_q}{a query built from csm_build_portfolio_query()}
}
\value{
the SQL query to build a CSM-like data frame
}
\description{
Get a CSM-like fundraiser total
}
