% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_target_asks.R
\name{csm_get_unit_target_ask_totals}
\alias{csm_get_unit_target_ask_totals}
\title{Get Unit Totals for Target Asks}
\usage{
csm_get_unit_target_ask_totals(unit_code, unit_desc)
}
\arguments{
\item{unit_code}{a unit code}

\item{unit_desc}{a unit description}
}
\value{
a one-line data frame with the unit and the unit's total of target asks
}
\description{
Get Unit Totals for Target Asks
}
