% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_more_details.R
\name{csm_maindata_details}
\alias{csm_maindata_details}
\title{CSM Maindata}
\usage{
csm_maindata_details(
  portfolio_q,
  start_dt = "2021-07-01",
  end_dt = "2022-06-30"
)
}
\arguments{
\item{portfolio_q}{a text string of a CSM portfolio query}

\item{start_dt}{the activity start date, as a string}

\item{end_dt}{the activity end date, as a string}
}
\value{
a text string that can pull the maindata CTE from the Campus Sumary Metrics
}
\description{
CSM Maindata
}
