% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_more_details.R
\name{csm_qualification_exclusions}
\alias{csm_qualification_exclusions}
\title{Plans Excluded from the CSM Qualifications}
\usage{
csm_qualification_exclusions(
  portfolio_q,
  start_dt = "2021-07-01",
  end_dt = "2022-06-30"
)
}
\arguments{
\item{portfolio_q}{a text string that queries the CDW for the portfolio, made by csm_build_portfolio_query()}

\item{start_dt}{the activity start date, as a string}

\item{end_dt}{the activity end date, as a string}
}
\value{
a text string that can be used to query the CDW for non-counted qualifications
}
\description{
the CTE that counts qualifications in the CSM has a complicated subquery that is MINUS'd from the list of qualifications. To help with debugging, this function builds the query to see what's in that subquery, to help you identify why plans might not be counted.
}
