% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm_qualifications.R
\name{csm_qualifications}
\alias{csm_qualifications}
\title{Qualifications Counted in the CSM}
\usage{
csm_qualifications(
  portfolio_q,
  office = "0",
  plan_type = "0",
  start_dt = "2021-07-01",
  end_dt = "2022-06-30"
)
}
\arguments{
\item{portfolio_q}{a text string that queries the CDW for the portfolio, made by csm_build_portfolio_query()}

\item{office}{the office code string. Must match whatever portfolio_q was built with}

\item{plan_type}{the play type code string. Must match whatever portfolio_q was built with}

\item{start_dt}{the activity start date, as a string}

\item{end_dt}{the activity end date, as a string}
}
\value{
a text string that can be used to query the CDW to get proposals to be counted as qualifications
}
\description{
Qualifications Counted in the CSM
}
