% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-to-str.R
\name{date_to_str}
\alias{date_to_str}
\title{Convert a Date to a String}
\usage{
date_to_str(x)
}
\arguments{
\item{x}{something that looks like a date (doesn't technically need to be an actual date)}
}
\value{
a string formatted in YYYY-MM-DD format
}
\description{
Convert a Date to a String
}
