% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dom-outcomes.R
\name{dom_outcomes}
\alias{dom_outcomes}
\title{DOM outcomes}
\usage{
dom_outcomes(
  FR_ID,
  Unit,
  Plan_Type = "0",
  BeginDate = "2021-07-01",
  EndDate = "2022-06-30"
)
}
\arguments{
\item{FR_ID}{fundraiser entity id}

\item{Unit}{fundraiser's unit code}

\item{Plan_Type}{plan type code. '0' for all plan types}

\item{BeginDate}{a string of a date in YYYY-MM-DD format}

\item{EndDate}{a string of a date in YYYY-MM-DD format}
}
\value{
a query to pull the DOM outcomes table. Or maybe the details? Note sure
}
\description{
not sure this has been tested
}
