% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contact_reports.R
\name{get_contact_reports}
\alias{get_contact_reports}
\title{Get Contacts Reports}
\usage{
get_contact_reports(date_start, date_end, by_fundraiser_id = NA, db = cdw)
}
\arguments{
\item{date_start}{the cutoff date, a string in YYYY-MM-DD format}

\item{date_end}{the cutoff date, a string in YYYY-MM-DD format}

\item{by_fundraiser_id}{a vector of fundraiser entity IDs. Use NA to pull all}

\item{db}{the database connection to use}
}
\value{
a database table of contact reports
}
\description{
Get Contacts Reports
}
