% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proposal_data.R
\name{get_proposal_data}
\alias{get_proposal_data}
\title{Get Proposal Data}
\usage{
get_proposal_data(primary_prop_only = T, primary_entity_only = T, db = cdw)
}
\arguments{
\item{primary_prop_only}{a logical determining whether you want all proposal rows or only the primary proposal row}

\item{primary_entity_only}{a logical determinng whether you want all entities or only the primary entity in each prospect}

\item{db}{the database conenction to use}
}
\value{
a database table with one row per assignment, proposal, prospect and entity. Inactive assignments, proposals and prospects are included
}
\description{
Get Proposal Data
}
