% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qualification_history.R
\name{get_qualifications_history}
\alias{get_qualifications_history}
\title{Get Qualification History}
\usage{
get_qualifications_history(db = cdw)
}
\arguments{
\item{db}{a connection to CDW2_PROD. Note that this will fail if you don't use PROD, b/c qualification stage history depends on history tables (which are only available in PROD)}
}
\value{
an tbl_Oracle of all qualifications, with two columns: PROPOSAL_ID, QUALIFICATION_DATE
}
\description{
Total qualifications are the first date where a proposal moves from QU in a positive direction. Positive stages for Total Qualifications include:
\itemize{
 \item{"CU"} {Cultivation}
 \item{"PD"} {Proposal Development}
 \item{"SP"} {Solicitation in Progress}
 \item{"GS"} {Gift Stewardship}
 \item{"DS"} {Donor Stewardship}
 \item{"RP"} {Referred to Other Program}
 \item{"DQ"} {Disqualified}
 \item{"TD"} {Turned Down}
}
}
