% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roles.R
\name{get_roles}
\alias{get_roles}
\title{Get Role Data}
\usage{
get_roles(role = "Major Gift", role_fy = 2021, db = cdw)
}
\arguments{
\item{role}{a vector of roles, ie. descriptions as shown in Tableau. Use NA to return all roles}

\item{role_fy}{a numeric, 4 digit fiscal year}

\item{db}{the database connections to use}
}
\value{
a database table (not collected) with one column: FUNDRAISER_ENTITY_ID
}
\description{
Get Role Data
}
