% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tab2_data.R
\name{get_tab2_data}
\alias{get_tab2_data}
\title{Get Data For Tab 2}
\usage{
get_tab2_data(
  collect = F,
  date_range_start = "2014-07-01",
  date_range_end = today(),
  unit = NA,
  fundraiser_ids = NA,
  role = NA,
  role_fy = currentFY,
  start_fy = 2015,
  db = cdw
)
}
\arguments{
\item{collect}{a logical - should the data be collected or should we return database queries}

\item{date_range_start}{a string (or Date, if you have to), preferably in YYYY-MM-DD format}

\item{date_range_end}{a string (or Date, if you have to), preferably in YYYY-MM-DD format}

\item{unit}{names of units to include}

\item{fundraiser_ids}{entity IDs of fundraisers to include}

\item{role}{roles to include}

\item{role_fy}{which fiscal year should we look at for roles}

\item{start_fy}{what fiscal year should the bar chart start at}

\item{db}{a valid database connection}
}
\value{
a list with 3 data frames, role_data, contact_data, and proposal_data
}
\description{
Get Data For Tab 2
}
