% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluck_captions.R
\name{make_caption}
\alias{make_caption}
\title{Make Caption}
\usage{
make_caption(single_result, ytd = F)
}
\arguments{
\item{single_result}{a single result from built_tab2(), ie. one of the list items that it returns}

\item{ytd}{a logical - should we should YTD (TRUE) or do we want the full FY  (FALSE)}
}
\value{
a column dataframe with type and val where type is the row name and val is the numbers displayed in the caption
}
\description{
make a single Tab2 caption
}
