% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fundraiser_name.R
\name{make_fundraiser_name}
\alias{make_fundraiser_name}
\title{Make PROSPECT_DEV style Fundraiser Names}
\usage{
make_fundraiser_name(db = cdw)
}
\arguments{
\item{db}{a valid CDW connection}
}
\value{
a two column tbl_sql: Entity ID and FUNDRAISER_NAME
}
\description{
Make PROSPECT_DEV style Fundraiser Names
}
