% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluck_captions.R
\name{pluck_captions}
\alias{pluck_captions}
\title{Pluck Caption}
\usage{
pluck_captions(tab2_results, ytd = T)
}
\arguments{
\item{tab2_results}{a list object built from built_tab2()}

\item{ytd}{a logical - should we should YTD (TRUE) or do we want the full FY  (FALSE)}
}
\value{
a list of dataframes, each with two columns, type (the rownames) and val (the numbers that should be displayed)
}
\description{
Build all the captions so you can test Tab 2
}
