% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluck_tab2_bar_numbers.R
\name{pluck_tab2_bar_numbers}
\alias{pluck_tab2_bar_numbers}
\title{Pluck Tab2 Bar Numbers}
\usage{
pluck_tab2_bar_numbers(tab2_results)
}
\arguments{
\item{tab2_results}{the results of build_tab2()}
}
\value{
a data frame with one line per section on Tab2 and one column per fiscal year
}
\description{
a simplified data frame to test all the bar numbers on Tab 2 against
}
