% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tab2.R
\name{build_tab2}
\alias{build_tab2}
\title{Build Tab 2}
\usage{
build_tab2(
  alldata,
  proposal_data,
  contact_data,
  qual_data,
  role_data,
  start_fy = 2015
)
}
\arguments{
\item{alldata}{a list of proposal, contact, qualification, and role data built by get_tab2_data().}

\item{proposal_data}{a data frame of proposal data, as generated by get_proposal_data (or the proposal section of get_tab2_data()). Only use this if you do not specify anything for alldata.}

\item{contact_data}{a data frame of contact report data, as generated by get_contact_reports (or the contact section get_tab2_data()). Only use this if you do not specify anything for alldata.}

\item{role_data}{a data frame of role data, as generated by get_roles (or the role section get_tab2_data()). Only use this if you do not specify anything for alldata.}

\item{start_fy}{what fiscal year should the bar charts start at}
}
\value{
a list with 8 elements, one for each of the metrics on Tab 2
}
\description{
Build Tab 2
}
